agents/coverage/requirements.txt
# Microsoft Agent Framework — core runtime
agent-framework>=0.1.0
# Self-hosted FastAPI + ASGI server (replaces azure-ai-agentserver)
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
# Azure OpenAI Responses API client
azure-ai-projects>=1.0.0
# Managed identity / DefaultAzureCredential
azure-identity>=1.15.0
# Custom MCP HTTP client (header injection for DeepSense CloudFront)
httpx>=0.27.0
# Environment variable loading
python-dotenv>=1.0.0
# Azure Monitor OpenTelemetry (optional — remove if not using App Insights)
azure-monitor-opentelemetry>=1.6.0
